Add a -diag option
Add -diag to build a pinfinder-debug.zip file containing information
about the backups on the machine, what pinfinder found out about them as
far as restrictions passcodes go and some other information that might
be useful for me to figure out what went wrong in situations where
pinfinder fails to work for someone.

It deliberately does not collect any personal information or the
contents of any files.  It does collect:

* The result of the pin search and any errors reported
* Operating system information
* Any restriction pins it finds successfully in backups
* The contents of the restrictions files if it can locate them
* A list of all the files in the backup (which isn't enough to generally
reveal even which applications are installed)